WIP moved gnmi-client-gen outside of model folder
