+++ models/testdevice-1.0.x/go.mod
@@ -13,3 +13,7 @@ require (
 	github.com/openconfig/ygot v0.14.0
 	google.golang.org/grpc v1.41.0
 )
+
+replace (
+	github.com/onosproject/config-models => ../../
+)
