+++ pkg/gnmi-client-gen/template/gnmi_client.go.tpl
